+++ CHANGELOG.md
@@ -1,5 +1,27 @@
-## [0.1.0] - 2025-08-13
+## [1.0.0] - 2025-08-13
 
 ### Added
 
  - Drone camera control
+
+## [1.0.1] - 2025-08-25
+
+### Added
+
+ - Drone marker animation
+ - multiplayer marker sync
+
+### Fixed
+
+ - some marker line glitch
+
+## [1.0.2] - 2025-08-22
+
+### Added
+
+ - Item marker in drone view
+ - Smoother drone move(i guess?)
+
+### Fixed
+
+ - A bug cause error in sync manager